Use hostname in list item instead of URL.
We also parse out the hostname on the client side before sending the
request to the API server (so we don't have to wait for the API server
to display the hostname in the pending list item). However, now we're
exposing / bundling two separate URI parsing libraries to the browser,
and that kinda sucks. Should be pretty easy to get rid of one (or pick
a different, smaller one that serves both purposes).